`columnType` can be `"number"` for `VirtualColumn`s
I'm not sure what is _SUPPOSED_ to happen here, but if the column is a virtual column then the `columnType` ended up being the string `"number"` rather than the `Number` constructor, and number values such as `"0"` were not converted.

I'm not sure if this should be unified upstream, so I added this quick fix.